rename files to match existing scheme
