+++ src/dsp/sony_s_dsp_gaussian.hpp
